requirements.txt
Django==4.2.7
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
stripe==7.6.0
python-dotenv==1.0.0
django-allauth==0.57.0
django-widget-tweaks==1.5.0
django-filter==23.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
mysqlclient==2.2.0 